Separate my term_animator
